requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
sqlalchemy>=2.0.40,<3.0.0
psycopg[binary]>=3.2.6,<4.0.0
psycopg2-binary>=2.9.9,<3.0.0
pydantic>=2.11.0,<3.0.0
pydantic-settings>=2.10.1,<3.0.0
python-dotenv>=1.1.0,<2.0.0
bcrypt>=5.0.0,<6.0.0
PyJWT>=2.10.1,<3.0.0
email-validator>=2.2.0,<3.0.0
httpx>=0.28.0,<1.0.0
